NCI-Agency/anet#50: Add all fields and adjust css
