--- client/src/pages/reports/PrintEmotion.css
@@ -1,24 +1,0 @@
-@page {
-  margin: 0;
-}
-@media print {
-  thead {
-    display: table-header-group;
-  }
-  tfoot {
-    display: table-footer-group;
-  }
-
-  button {
-    display: none;
-  }
-
-  body {
-    margin: 0;
-  }
-  /* messes with print pdf*/
-  svg > text#__react_svg_text_measurement_id {
-    display: none;
-  }
-
-}
